Delete, union, and rename created.
Also, I haven't been using the table type, as I've just referred to each table with its root node

The Table type seems redundant given that the root node contains the entire table and all the necessary functions for manipulating it.
This is a problem for the functions returning table types.