Fleshed out the comparisons for condition's evaluate function.
I had to pass the TableRootNode in addition to the row being compared in order to make it work (I'm trash when it comes to handling private vs public classes.  I'm sure there's a better approach than this, but it should work, and it should be something that can be handled by only changing select and other methods that use evaluate, since these all have a string telling them which table they're taking data from).