Pseudo code for Typecasting Query created.
Once we have a better grasp on how we're handling the SQL command creation, we can finish this up.
I was a bit iffy on how the for each for the hash maps works, but that shouldn't impact the rest of the logic, and I was sure to comment a lot on what each line is supposed to be doing, so it shouldn't be hard to fix the syntax.